Move Monitor and Plot to their own file
This moves these classes from mission_basic to their own submodule,
trajectory.Monitor. This makes the main file cleaner and keeps a modular
design.
